% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluecodes.R
\name{generate_modelfit}
\alias{generate_modelfit}
\title{Function to return a glue code function to call raschmodel, plmodel, gpcmodel or rsmodel,}
\usage{
generate_modelfit(
  start = NULL,
  weights = NULL,
  estfun = FALSE,
  object = FALSE,
  itemtype = c("Rasch", "2PL", "3PL", "3PLu", "4PL", "1PL", "RM", "GPCM", "PCM", "RSM"),
  grouppars = FALSE,
  vcov = TRUE,
  method = "BFGS",
  maxit = 500L,
  reltol = 1e-10,
  deriv = "sum",
  hessian = TRUE,
  full = TRUE
)
}
\arguments{
\item{start}{an optional vector or list of starting values (see \code{\link[psychotools]{raschmodel}}, \code{\link[psychotools]{plmodel}}, \code{\link[psychotools]{gpcmodel}}, or \code{\link[psychotools]{rsmodel}}).}

\item{weights}{an optional vector of weights (interpreted as case weights).}

\item{estfun}{a (potentially user-defined) estfun function.}

\item{object}{a (potentially user-defined) model object.}

\item{itemtype}{character string, specifying the type of IRT model to be estimated (see details below)}

\item{grouppars}{logical. Should the estimated distributional group parameters of a multiple group model be included in the model parameters? see plmodel and gpcmodel. I think I can remove this parameter?}

\item{vcov}{logical or character specifying the type of variance-covariance matrix (if any) computed for the final models when fitted using MML (see \code{\link[psychotools]{plmodel}} and \code{\link[psychotools]{gpcmodel}}).}

\item{method}{control parameter for the optimizer employed by \code{\link[mirt]{mirt}} for the EM algorithm when models are fitted using MML (see \code{\link[psychotools]{plmodel}} and \code{\link[psychotools]{gpcmodel}}).}

\item{maxit}{control parameter for the optimizer employed by (see \code{\link[psychotools]{raschmodel}}, \code{\link[psychotools]{plmodel}}, \code{\link[psychotools]{gpcmodel}}, or \code{\link[psychotools]{rsmodel}})}

\item{reltol}{control parameter for the optimizer employed by (see \code{\link[psychotools]{raschmodel}}, \code{\link[psychotools]{plmodel}}, \code{\link[psychotools]{gpcmodel}}, or \code{\link[psychotools]{rsmodel}})}

\item{deriv}{character. Which type of derivatives should be used for computing gradient and Hessian matrix when fitting using \code{\link[psychotools]{raschmodel}} or \code{\link[psychotools]{rsmodel}}?}

\item{hessian}{logical. Should the Hessian be computed for models fitted by \code{\link[psychotools]{raschmodel}} or \code{\link[psychotools]{rsmodel}}?}

\item{full}{logical. Should a full model object be returned for models fitted by \code{\link[psychotools]{raschmodel}} or \code{\link[psychotools]{rsmodel}}?}
}
\value{
A glue code function with parameters y, x
}
\description{
Function to return a glue code function to call raschmodel, plmodel, gpcmodel or rsmodel,
}
